+++ spec/support/shared_contexts/cache_store.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+shared_context 'cache store' do
+  after { cache_store.clear }
+
+  let(:cache_store) { build_cache_store }
+end